Add Game Lobby Menu
The lobby is openned by clicking "new game" in the main menu.
Currently only two start dates are in the listbox,
but the intention is for saved games to be present as well.
Most of the space on the right-hand panel is reserved for
information about the selected country and other players.
The purpose of the separation is to allow a map to be rendered
behind the menu.
The start button can be pressed only after a save has been selected.